+++ resources/views/examples/date-picker/min-max.blade.php
@@ -1,0 +1,3 @@
+{{-- min/max are date bounds. Default out-of-range="disable": dates outside [10–20 Jun] are
+     struck-through and not selectable. --}}
+<x-ui.date-picker name="d" min="2026-06-10" max="2026-06-20" default-month="2026-06-10" />